Code pushed

+++ src/model/messageModel/md_messages_templates.js
@@ -2,8 +2,8 @@ import db from "../../config/db.js";
 
 const createTableQuery = `
 CREATE TABLE IF NOT EXISTS md_message_templates (
-  id INT AUTO_INCREMENT PRIMARY KEY,
-  template_code INT NOT NULL,
+  id BIGINT AUTO_INCREMENT PRIMARY KEY,
+  template_code BIGINT NOT NULL,
   language_id VARCHAR(3) NOT NULL,
   channel SMALLINT,
   subject VARCHAR(500) NOT NULL,
